+++ src/components/Text/Text.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { css, cx } from '@emotion/css';
 import { DataFrame } from '@grafana/data';
 import { Alert, useTheme2 } from '@grafana/ui';
 import { generateHtml } from '../../helpers';
@@ -32,7 +33,18 @@ export const Text: React.FC<Props> = ({ options, frame }) => {
    * Theme
    */
   const theme = useTheme2();
-  const styles = getStyles(theme, options.styles);
+  const styles = getStyles(theme);
+
+  /**
+   * Styles
+   */
+  const className = cx(
+    styles.highlight,
+    styles.frame,
+    css`
+      ${options.styles}
+    `
+  );
 
   try {
     /**
@@ -41,7 +53,7 @@ export const Text: React.FC<Props> = ({ options, frame }) => {
     if (!frame?.length) {
       return (
         <div
-          className={styles.frame}
+          className={className}
           dangerouslySetInnerHTML={{ __html: generateHtml({}, options.defaultContent, options.helpers) }}
         />
       );
@@ -67,7 +79,7 @@ export const Text: React.FC<Props> = ({ options, frame }) => {
           {data.map((row, key) => (
             <div
               key={key}
-              className={styles.frame}
+              className={className}
               dangerouslySetInnerHTML={{ __html: generateHtml(row, options.content, options.helpers) }}
             />
           ))}
@@ -80,7 +92,7 @@ export const Text: React.FC<Props> = ({ options, frame }) => {
      */
     return (
       <div
-        className={styles.frame}
+        className={className}
         dangerouslySetInnerHTML={{ __html: generateHtml({ data }, options.content, options.helpers) }}
       />
     );