+++ src/constants/default.ts
@@ -5,7 +5,7 @@ import { CodeLanguage, Format } from './editor';
  * Default Options
  */
 export const DefaultOptions: TextOptions = {
-  content: '{{json @root}}',
+  content: '```json\n{{{json @root}}}\n```',
   defaultContent: "The query didn't return any results.",
   editors: [],
   editor: { height: 200, format: Format.AUTO, language: CodeLanguage.MARKDOWN },